Study aliasing 04-Field-aliasing...
I have learnt and summarize how to use aliasing in serialization and
deserealization. I learnt to use model_dump and model_dump_json with the
parameter by_alias: bool. Besides, I learnt how to customize a
serializar and use an special object to recognize if we are calling
model_dump or model_dump_json. Last but not list, I use alias_generator
for the model fields.